Alpha version
Level One of the game is almost done. The game is having a starting splash scene and it proceeds to the first level scene. Each scene is packed with audio efects. One player from the first level will be eliminated and rest three will be proceeded to the second level.